backend/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.2
    # via
    #   django
    #   djoser
attrs==21.4.0
    # via pytest
certifi==2022.6.15
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.0
    # via requests
coreapi==2.3.3
    # via djoser
coreschema==0.0.4
    # via coreapi
cryptography==37.0.4
    # via social-auth-core
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
django==3.2.14
    # via
    #   -r requirements.in
    #   django-debug-toolbar
    #   djangohexadecimalfield
    #   djangorestframework
    #   djangorestframework-simplejwt
django-debug-toolbar==3.5.0
    # via -r requirements.in
django-templated-mail==1.1.1
    # via djoser
djangohexadecimalfield==0.0.3
    # via -r requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==4.8.0
    # via
    #   -r requirements.in
    #   djoser
djoser==2.1.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
idna==3.3
    # via requests
iniconfig==1.1.1
    # via pytest
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
markupsafe==2.1.1
    # via jinja2
oauthlib==3.2.0
    # via
    #   requests-oauthlib
    #   social-auth-core
packaging==21.3
    # via pytest
pillow==9.2.0
    # via -r requirements.in
pluggy==1.0.0
    # via pytest
psycopg2-binary==2.9.3
    # via -r requirements.in
py==1.11.0
    # via pytest
pycparser==2.21
    # via cffi
pyjwt==2.4.0
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
pyparsing==3.0.9
    # via packaging
pytest==7.1.2
    # via
    #   -r requirements.in
    #   pytest-django
pytest-django==4.5.2
    # via -r requirements.in
python-dotenv==0.20.0
    # via -r requirements.in
python3-openid==3.2.0
    # via social-auth-core
pytz==2022.1
    # via
    #   django
    #   djangorestframework
requests==2.28.1
    # via
    #   coreapi
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.1
    # via social-auth-core
six==1.16.0
    # via social-auth-app-django
social-auth-app-django==4.0.0
    # via djoser
social-auth-core==4.3.0
    # via social-auth-app-django
sorl-thumbnail==12.8.0
    # via -r requirements.in
sqlparse==0.4.2
    # via
    #   django
    #   django-debug-toolbar
tomli==2.0.1
    # via pytest
uritemplate==4.1.1
    # via coreapi
urllib3==1.26.11
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
